+++ socketio/Cargo.toml
@@ -7,7 +7,11 @@ description = "An implementation of a socketio client written in rust."
 readme = "../README.md"
 repository = "https://github.com/1c3t3a/rust-socketio"
 keywords = ["socketio", "engineio", "network", "protocol", "client"]
-categories = ["network-programming", "web-programming", "web-programming::websocket"]
+categories = [
+    "network-programming",
+    "web-programming",
+    "web-programming::websocket",
+]
 license = "MIT"
 
 [dependencies]
@@ -22,10 +26,13 @@ thiserror = "1.0"
 native-tls = "0.2.11"
 url = "2.3.1"
 tokio = { version = "1.16.1", optional = true }
-futures-util = { version = "0.3", default-features = false, features = ["sink"], optional = true }
+futures-util = { version = "0.3", default-features = false, features = [
+    "sink",
+], optional = true }
 async-stream = { version = "0.3.5", optional = true }
 log = "0.4.17"
 serde = "1.0.164"
+async-trait = "0.1.73"
 
 [dev-dependencies]
 cargo-tarpaulin = "0.18.5"
@@ -38,7 +45,13 @@ features = ["macros"]
 [features]
 default = []
 async-callbacks = ["rust_engineio/async-callbacks"]
-async = ["async-callbacks", "rust_engineio/async", "tokio", "futures-util", "async-stream"]
+async = [
+    "async-callbacks",
+    "rust_engineio/async",
+    "tokio",
+    "futures-util",
+    "async-stream",
+]
 
 [[example]]
 name = "async"
