+++ socketio/src/asynchronous/client/mod.rs
@@ -2,4 +2,7 @@ mod ack;
 pub(crate) mod builder;
 #[cfg(feature = "async-callbacks")]
 mod callback;
+mod raw_client;
+
+pub use raw_client::RawClient;
 pub(crate) mod client;
