+++ socketio/src/asynchronous/client/raw_client.rs
@@ -1,0 +1,888 @@
+use std::{collections::HashMap, ops::DerefMut, pin::Pin, sync::Arc};
+
+use futures_util::{future::BoxFuture, stream, Stream, StreamExt};
+use log::trace;
+use rand::{thread_rng, Rng};
+use serde_json::Value;
+use tokio::{
+    sync::RwLock,
+    time::{Duration, Instant},
+};
+
+use super::callback::{Callback, DynAsyncAnyCallback, DynAsyncCallback};
+use crate::{
+    asynchronous::socket::Socket as InnerSocket,
+    error::{Error, Result},
+    packet::{Packet, PacketId},
+    Event, Payload,
+};
+
+/// Represents an `Ack` as given back to the caller. Holds the internal `id` as
+/// well as the current ack'ed state. Holds data which will be accessible as
+/// soon as the ack'ed state is set to true. An `Ack` that didn't get ack'ed
+/// won't contain data.
+#[derive(Debug)]
+pub(crate) struct Ack {
+    pub id: i32,
+    pub timeout: Duration,
+    pub time_started: Instant,
+    pub callback: Callback<DynAsyncCallback>,
+}
+
+/// A socket which handles communication with the server. It's initialized with
+/// a specific address as well as an optional namespace to connect to. If `None`
+/// is given the server will connect to the default namespace `"/"`.
+#[derive(Clone)]
+pub struct RawClient {
+    /// The inner socket client to delegate the methods to.
+    socket: InnerSocket,
+    on: Arc<RwLock<HashMap<Event, Callback<DynAsyncCallback>>>>,
+    on_any: Arc<RwLock<Option<Callback<DynAsyncAnyCallback>>>>,
+    outstanding_acks: Arc<RwLock<Vec<Ack>>>,
+    // namespace, for multiplexing messages
+    nsp: String,
+    // Data send in the opening packet (commonly used as for auth)
+    auth: Option<serde_json::Value>,
+}
+
+impl RawClient {
+    /// Creates a socket with a certain address to connect to as well as a
+    /// namespace. If `None` is passed in as namespace, the default namespace
+    /// `"/"` is taken.
+    /// ```
+    pub(crate) fn new<T: Into<String>>(
+        socket: InnerSocket,
+        namespace: T,
+        on: Arc<RwLock<HashMap<Event, Callback<DynAsyncCallback>>>>,
+        on_any: Arc<RwLock<Option<Callback<DynAsyncAnyCallback>>>>,
+        auth: Option<serde_json::Value>,
+    ) -> Result<Self> {
+        Ok(Self {
+            socket,
+            nsp: namespace.into(),
+            on,
+            on_any,
+            outstanding_acks: Arc::new(RwLock::new(Vec::new())),
+            auth,
+        })
+    }
+
+    /// Connects the client to a server. Afterwards the `emit_*` methods can be
+    /// called to interact with the server.
+    pub(crate) async fn connect(&self) -> Result<()> {
+        // Connect the underlying socket
+        self.socket.connect().await?;
+
+        // construct the opening packet
+        let auth = self.auth.as_ref().map(|data| data.to_string());
+        let open_packet = Packet::new(PacketId::Connect, self.nsp.clone(), auth, None, 0, None);
+
+        self.socket.send(open_packet).await?;
+
+        Ok(())
+    }
+
+    /// Sends a message to the server using the underlying `engine.io` protocol.
+    /// This message takes an event, which could either be one of the common
+    /// events like "message" or "error" or a custom event like "foo". But be
+    /// careful, the data string needs to be valid JSON. It's recommended to use
+    /// a library like `serde_json` to serialize the data properly.
+    ///
+    /// # Example
+    /// ```
+    /// use rust_socketio::{asynchronous::{ClientBuilder, Client}, Payload};
+    /// use serde_json::json;
+    /// use futures_util::FutureExt;
+    ///
+    /// #[tokio::main]
+    /// async fn main() {
+    ///     let mut socket = ClientBuilder::new("http://localhost:4200/")
+    ///         .on("test", |payload: Payload, socket: Client| {
+    ///             async move {
+    ///                 println!("Received: {:#?}", payload);
+    ///                 socket.emit("test", json!({"hello": true})).await.expect("Server unreachable");
+    ///             }.boxed()
+    ///         })
+    ///         .connect()
+    ///         .await
+    ///         .expect("connection failed");
+    ///
+    ///     let json_payload = json!({"token": 123});
+    ///
+    ///     let result = socket.emit("foo", json_payload).await;
+    ///
+    ///     assert!(result.is_ok());
+    /// }
+    /// ```
+    #[inline]
+    pub async fn emit<E, D>(&self, event: E, data: D) -> Result<()>
+    where
+        E: Into<Event>,
+        D: Into<Payload>,
+    {
+        self.socket.emit(&self.nsp, event.into(), data.into()).await
+    }
+
+    /// Sends a message to the server but `alloc`s an `ack` to check whether the
+    /// server responded in a given time span. This message takes an event, which
+    /// could either be one of the common events like "message" or "error" or a
+    /// custom event like "foo", as well as a data parameter. But be careful,
+    /// in case you send a [`Payload::String`], the string needs to be valid JSON.
+    /// It's even recommended to use a library like serde_json to serialize the data properly.
+    /// It also requires a timeout `Duration` in which the client needs to answer.
+    /// If the ack is acked in the correct time span, the specified callback is
+    /// called. The callback consumes a [`Payload`] which represents the data send
+    /// by the server.
+    ///
+    /// Please note that the requirements on the provided callbacks are similar to the ones
+    /// for [`crate::asynchronous::ClientBuilder::on`].
+    /// # Example
+    /// ```
+    /// use rust_socketio::{asynchronous::{ClientBuilder, Client}, Payload};
+    /// use serde_json::json;
+    /// use std::time::Duration;
+    /// use std::thread::sleep;
+    /// use futures_util::FutureExt;
+    ///
+    /// #[tokio::main]
+    /// async fn main() {
+    ///     let mut socket = ClientBuilder::new("http://localhost:4200/")
+    ///         .on("foo", |payload: Payload, _| async move { println!("Received: {:#?}", payload) }.boxed())
+    ///         .connect()
+    ///         .await
+    ///         .expect("connection failed");
+    ///
+    ///     let ack_callback = |message: Payload, socket: Client| {
+    ///         async move {
+    ///             match message {
+    ///                 Payload::String(str) => println!("{}", str),
+    ///                 Payload::Binary(bytes) => println!("Received bytes: {:#?}", bytes),
+    ///             }
+    ///         }.boxed()
+    ///     };    
+    ///
+    ///
+    ///     let payload = json!({"token": 123});
+    ///     socket.emit_with_ack("foo", payload, Duration::from_secs(2), ack_callback).await.unwrap();
+    ///
+    ///     sleep(Duration::from_secs(2));
+    /// }
+    /// ```
+    #[inline]
+    pub async fn emit_with_ack<F, E, D>(
+        &self,
+        event: E,
+        data: D,
+        timeout: Duration,
+        callback: F,
+    ) -> Result<()>
+    where
+        F: for<'a> std::ops::FnMut(Payload, RawClient) -> BoxFuture<'static, ()>
+            + 'static
+            + Send
+            + Sync,
+        E: Into<Event>,
+        D: Into<Payload>,
+    {
+        let id = thread_rng().gen_range(0..999);
+        let socket_packet =
+            self.socket
+                .build_packet_for_payload(data.into(), event.into(), &self.nsp, Some(id))?;
+
+        let ack = Ack {
+            id,
+            time_started: Instant::now(),
+            timeout,
+            callback: Callback::<DynAsyncCallback>::new(callback),
+        };
+
+        // add the ack to the tuple of outstanding acks
+        self.outstanding_acks.write().await.push(ack);
+
+        self.socket.send(socket_packet).await?;
+        Ok(())
+    }
+
+    /// Disconnects this client from the server by sending a `socket.io` closing
+    /// packet.
+    /// # Example
+    /// ```rust
+    /// use rust_socketio::{asynchronous::{ClientBuilder, Client}, Payload};
+    /// use serde_json::json;
+    /// use futures_util::{FutureExt, future::BoxFuture};
+    ///
+    /// #[tokio::main]
+    /// async fn main() {
+    ///     // apparently the syntax for functions is a bit verbose as rust currently doesn't
+    ///     // support an `AsyncFnMut` type that conform with async functions
+    ///     fn handle_test(payload: Payload, socket: Client) -> BoxFuture<'static, ()> {
+    ///         async move {
+    ///             println!("Received: {:#?}", payload);
+    ///             socket.emit("test", json!({"hello": true})).await.expect("Server unreachable");
+    ///         }.boxed()
+    ///     }
+    ///
+    ///     let mut socket = ClientBuilder::new("http://localhost:4200/")
+    ///         .on("test", handle_test)
+    ///         .connect()
+    ///         .await
+    ///         .expect("connection failed");
+    ///
+    ///     let json_payload = json!({"token": 123});
+    ///
+    ///     socket.emit("foo", json_payload).await;
+    ///
+    ///     // disconnect from the server
+    ///     socket.disconnect().await;
+    /// }
+    /// ```
+    pub async fn disconnect(&self) -> Result<()> {
+        let disconnect_packet =
+            Packet::new(PacketId::Disconnect, self.nsp.clone(), None, None, 0, None);
+
+        self.socket.send(disconnect_packet).await?;
+        self.socket.disconnect().await?;
+        self.callback(&Event::Close, "").await?;
+        Ok(())
+    }
+
+    async fn callback<P: Into<Payload>>(&self, event: &Event, payload: P) -> Result<()> {
+        let mut on = self.on.write().await;
+        let mut on_any = self.on_any.write().await;
+
+        let on_lock = on.deref_mut();
+        let on_any_lock = on_any.deref_mut();
+        let payload = payload.into();
+
+        if let Some(callback) = on_lock.get_mut(event) {
+            callback(payload.clone(), self.clone()).await;
+        }
+
+        // Call on_any for all common and custom events.
+        match event {
+            Event::Message | Event::Custom(_) => {
+                if let Some(callback) = on_any_lock {
+                    callback(event.clone(), payload, self.clone()).await;
+                }
+            }
+            _ => (),
+        }
+
+        drop(on);
+        drop(on_any);
+        Ok(())
+    }
+
+    /// Handles the incoming acks and classifies what callbacks to call and how.
+    #[inline]
+    async fn handle_ack(&self, socket_packet: &Packet) -> Result<()> {
+        let mut to_be_removed = Vec::new();
+        if let Some(id) = socket_packet.id {
+            for (index, ack) in self.outstanding_acks.write().await.iter_mut().enumerate() {
+                if ack.id == id {
+                    to_be_removed.push(index);
+
+                    if ack.time_started.elapsed() < ack.timeout {
+                        if let Some(ref payload) = socket_packet.data {
+                            ack.callback.deref_mut()(
+                                Payload::String(payload.to_owned()),
+                                self.clone(),
+                            )
+                            .await;
+                        }
+                        if let Some(ref attachments) = socket_packet.attachments {
+                            if let Some(payload) = attachments.get(0) {
+                                ack.callback.deref_mut()(
+                                    Payload::Binary(payload.to_owned()),
+                                    self.clone(),
+                                )
+                                .await;
+                            }
+                        }
+                    } else {
+                        trace!("Received an Ack that is now timed out (elapsed time was longer than specified duration)");
+                    }
+                }
+            }
+            for index in to_be_removed {
+                self.outstanding_acks.write().await.remove(index);
+            }
+        }
+        Ok(())
+    }
+
+    /// Handles a binary event.
+    #[inline]
+    async fn handle_binary_event(&self, packet: &Packet) -> Result<()> {
+        let event = if let Some(string_data) = &packet.data {
+            string_data.replace('\"', "").into()
+        } else {
+            Event::Message
+        };
+
+        if let Some(attachments) = &packet.attachments {
+            if let Some(binary_payload) = attachments.get(0) {
+                self.callback(&event, Payload::Binary(binary_payload.to_owned()))
+                    .await?;
+            }
+        }
+        Ok(())
+    }
+
+    /// A method that parses a packet and eventually calls the corresponding
+    /// callback with the supplied data.
+    async fn handle_event(&self, packet: &Packet) -> Result<()> {
+        let Some(ref data) = packet.data else {
+            return Ok(());
+        };
+
+        // a socketio message always comes in one of the following two flavors (both JSON):
+        // 1: `["event", "msg"]`
+        // 2: `["msg"]`
+        // in case 2, the message is ment for the default message event, in case 1 the event
+        // is specified
+        if let Ok(Value::Array(contents)) = serde_json::from_str::<Value>(data) {
+            let (event, data) = if contents.len() > 1 {
+                // case 1
+                let event = match contents.first() {
+                    Some(Value::String(ev)) => Event::from(ev.as_str()),
+                    _ => Event::Message,
+                };
+
+                (event, contents.get(1).ok_or(Error::IncompletePacket())?)
+            } else {
+                // case 2
+                (
+                    Event::Message,
+                    contents.first().ok_or(Error::IncompletePacket())?,
+                )
+            };
+
+            // call the correct callback
+            self.callback(&event, data.to_string()).await?;
+        }
+
+        Ok(())
+    }
+
+    /// Handles the incoming messages and classifies what callbacks to call and how.
+    /// This method is later registered as the callback for the `on_data` event of the
+    /// engineio client.
+    #[inline]
+    async fn handle_socketio_packet(&self, packet: &Packet) -> Result<()> {
+        if packet.nsp == self.nsp {
+            match packet.packet_type {
+                PacketId::Ack | PacketId::BinaryAck => {
+                    if let Err(err) = self.handle_ack(packet).await {
+                        self.callback(&Event::Error, err.to_string()).await?;
+                        return Err(err);
+                    }
+                }
+                PacketId::BinaryEvent => {
+                    if let Err(err) = self.handle_binary_event(packet).await {
+                        self.callback(&Event::Error, err.to_string()).await?;
+                    }
+                }
+                PacketId::Connect => {
+                    self.callback(&Event::Connect, "").await?;
+                }
+                PacketId::Disconnect => {
+                    self.callback(&Event::Close, "").await?;
+                }
+                PacketId::ConnectError => {
+                    self.callback(
+                        &Event::Error,
+                        String::from("Received an ConnectError frame: ")
+                            + packet
+                                .data
+                                .as_ref()
+                                .unwrap_or(&String::from("\"No error message provided\"")),
+                    )
+                    .await?;
+                }
+                PacketId::Event => {
+                    if let Err(err) = self.handle_event(packet).await {
+                        self.callback(&Event::Error, err.to_string()).await?;
+                    }
+                }
+            }
+        }
+        Ok(())
+    }
+
+    /// Returns the packet stream for the client.
+    pub(crate) fn as_stream<'a>(
+        &'a self,
+    ) -> Pin<Box<dyn Stream<Item = Result<Packet>> + Send + 'a>> {
+        stream::unfold(self.socket.clone(), |mut socket| async {
+            // wait for the next payload
+            let packet: Option<std::result::Result<Packet, Error>> = socket.next().await;
+            match packet {
+                // end the stream if the underlying one is closed
+                None => None,
+                Some(Err(err)) => {
+                    // call the error callback
+                    match self.callback(&Event::Error, err.to_string()).await {
+                        Err(callback_err) => Some((Err(callback_err), socket)),
+                        Ok(_) => Some((Err(err), socket)),
+                    }
+                }
+                Some(Ok(packet)) => match self.handle_socketio_packet(&packet).await {
+                    Err(callback_err) => Some((Err(callback_err), socket)),
+                    Ok(_) => Some((Ok(packet), socket)),
+                },
+            }
+        })
+        .boxed()
+    }
+}
+
+#[cfg(test)]
+mod test {
+
+    use std::{sync::Arc, time::Duration};
+
+    use bytes::Bytes;
+    use futures_util::{FutureExt, StreamExt};
+    use native_tls::TlsConnector;
+    use serde_json::json;
+    use tokio::{
+        sync::mpsc,
+        time::{sleep, timeout},
+    };
+
+    use crate::{
+        asynchronous::client::{builder::ClientBuilder, RawClient},
+        error::Result,
+        packet::{Packet, PacketId},
+        Payload, TransportType,
+    };
+
+    #[tokio::test]
+    async fn socket_io_integration() -> Result<()> {
+        let url = crate::test::socket_io_server();
+
+        let socket = ClientBuilder::new(url)
+            .on("test", |msg, _| {
+                async {
+                    match msg {
+                        Payload::String(str) => println!("Received string: {}", str),
+                        Payload::Binary(bin) => println!("Received binary data: {:#?}", bin),
+                    }
+                }
+                .boxed()
+            })
+            .connect()
+            .await?;
+
+        let payload = json!({"token": 123_i32});
+        let result = socket
+            .emit("test", Payload::String(payload.to_string()))
+            .await;
+
+        assert!(result.is_ok());
+
+        let ack_callback = |message: Payload, socket: RawClient| {
+            async move {
+                let result = socket
+                    .emit(
+                        "test",
+                        Payload::String(json!({"got ack": true}).to_string()),
+                    )
+                    .await;
+                assert!(result.is_ok());
+                println!("Yehaa! My ack got acked?");
+                if let Payload::String(str) = message {
+                    println!("Received string Ack");
+                    println!("Ack data: {}", str);
+                }
+            }
+            .boxed()
+        };
+
+        let ack = socket
+            .emit_with_ack(
+                "test",
+                Payload::String(payload.to_string()),
+                Duration::from_secs(1),
+                ack_callback,
+            )
+            .await;
+        assert!(ack.is_ok());
+
+        sleep(Duration::from_secs(2)).await;
+
+        assert!(socket.disconnect().await.is_ok());
+
+        Ok(())
+    }
+
+    #[tokio::test]
+    async fn socket_io_async_callback() -> Result<()> {
+        // Test whether asynchronous callbacks are fully executed.
+        let url = crate::test::socket_io_server();
+
+        // This synchronization mechanism is used to let the test know that the end of the
+        // async callback was reached.
+        let notify = Arc::new(tokio::sync::Notify::new());
+        let notify_clone = notify.clone();
+
+        let socket = ClientBuilder::new(url)
+            .on("test", move |_, _| {
+                let cl = notify_clone.clone();
+                async move {
+                    sleep(Duration::from_secs(1)).await;
+                    // The async callback should be awaited and not aborted.
+                    // Thus, the notification should be called.
+                    cl.notify_one();
+                }
+                .boxed()
+            })
+            .connect()
+            .await?;
+
+        let payload = json!({"token": 123_i32});
+        let result = socket
+            .emit("test", Payload::String(payload.to_string()))
+            .await;
+
+        assert!(result.is_ok());
+        // If the timeout did not trigger, the async callback was fully executed.
+        let timeout = timeout(Duration::from_secs(5), notify.notified()).await;
+        assert!(timeout.is_ok());
+
+        Ok(())
+    }
+
+    #[tokio::test]
+    async fn socket_io_builder_integration() -> Result<()> {
+        let url = crate::test::socket_io_server();
+
+        // test socket build logic
+        let socket_builder = ClientBuilder::new(url);
+
+        let tls_connector = TlsConnector::builder()
+            .use_sni(true)
+            .build()
+            .expect("Found illegal configuration");
+
+        let socket = socket_builder
+            .namespace("/admin")
+            .tls_config(tls_connector)
+            .opening_header("accept-encoding", "application/json")
+            .on("test", |str, _| {
+                async move { println!("Received: {:#?}", str) }.boxed()
+            })
+            .on("message", |payload, _| {
+                async move { println!("{:#?}", payload) }.boxed()
+            })
+            .connect()
+            .await?;
+
+        assert!(socket.emit("message", json!("Hello World")).await.is_ok());
+
+        assert!(socket
+            .emit("binary", Bytes::from_static(&[46, 88]))
+            .await
+            .is_ok());
+
+        assert!(socket
+            .emit_with_ack(
+                "binary",
+                json!("pls ack"),
+                Duration::from_secs(1),
+                |payload, _| async move {
+                    println!("Yehaa the ack got acked");
+                    println!("With data: {:#?}", payload);
+                }
+                .boxed()
+            )
+            .await
+            .is_ok());
+
+        sleep(Duration::from_secs(2)).await;
+
+        Ok(())
+    }
+
+    #[tokio::test]
+    async fn socket_io_builder_integration_iterator() -> Result<()> {
+        let url = crate::test::socket_io_server();
+
+        // test socket build logic
+        let socket_builder = ClientBuilder::new(url);
+
+        let tls_connector = TlsConnector::builder()
+            .use_sni(true)
+            .build()
+            .expect("Found illegal configuration");
+
+        let socket = socket_builder
+            .namespace("/admin")
+            .tls_config(tls_connector)
+            .opening_header("accept-encoding", "application/json")
+            .on("test", |str, _| {
+                async move { println!("Received: {:#?}", str) }.boxed()
+            })
+            .on("message", |payload, _| {
+                async move { println!("{:#?}", payload) }.boxed()
+            })
+            .connect_raw()
+            .await?;
+
+        assert!(socket.emit("message", json!("Hello World")).await.is_ok());
+
+        assert!(socket
+            .emit("binary", Bytes::from_static(&[46, 88]))
+            .await
+            .is_ok());
+
+        assert!(socket
+            .emit_with_ack(
+                "binary",
+                json!("pls ack"),
+                Duration::from_secs(1),
+                |payload, _| async move {
+                    println!("Yehaa the ack got acked");
+                    println!("With data: {:#?}", payload);
+                }
+                .boxed()
+            )
+            .await
+            .is_ok());
+
+        test_socketio_socket(socket, "/admin".to_owned()).await
+    }
+
+    #[tokio::test]
+    async fn socket_io_on_any_integration() -> Result<()> {
+        let url = crate::test::socket_io_server();
+
+        let (tx, mut rx) = mpsc::channel(2);
+
+        let mut _socket = ClientBuilder::new(url)
+            .namespace("/")
+            .auth(json!({ "password": "123" }))
+            .on_any(move |event, payload, _| {
+                let clone_tx = tx.clone();
+                async move {
+                    if let Payload::String(str) = payload {
+                        println!("{event}: {str}");
+                    }
+                    clone_tx.send(String::from(event)).await.unwrap();
+                }
+                .boxed()
+            })
+            .connect()
+            .await?;
+
+        let event = rx.recv().await.unwrap();
+        assert_eq!(event, "message");
+
+        let event = rx.recv().await.unwrap();
+        assert_eq!(event, "test");
+
+        Ok(())
+    }
+
+    #[tokio::test]
+    async fn socket_io_auth_builder_integration() -> Result<()> {
+        let url = crate::test::socket_io_auth_server();
+        let nsp = String::from("/admin");
+        let socket = ClientBuilder::new(url)
+            .namespace(nsp.clone())
+            .auth(json!({ "password": "123" }))
+            .connect_raw()
+            .await?;
+
+        // open packet
+        let mut socket_stream = socket.as_stream();
+        let _ = socket_stream.next().await.unwrap()?;
+
+        println!("Here12");
+        let packet = socket_stream.next().await.unwrap()?;
+        assert_eq!(
+            packet,
+            Packet::new(
+                PacketId::Event,
+                nsp,
+                Some("[\"auth\",\"success\"]".to_owned()),
+                None,
+                0,
+                None
+            )
+        );
+
+        Ok(())
+    }
+
+    #[tokio::test]
+    async fn socketio_polling_integration() -> Result<()> {
+        let url = crate::test::socket_io_server();
+        let socket = ClientBuilder::new(url.clone())
+            .transport_type(TransportType::Polling)
+            .connect_raw()
+            .await?;
+        test_socketio_socket(socket, "/".to_owned()).await
+    }
+
+    #[tokio::test]
+    async fn socket_io_websocket_integration() -> Result<()> {
+        let url = crate::test::socket_io_server();
+        let socket = ClientBuilder::new(url.clone())
+            .transport_type(TransportType::Websocket)
+            .connect_raw()
+            .await?;
+        test_socketio_socket(socket, "/".to_owned()).await
+    }
+
+    #[tokio::test]
+    async fn socket_io_websocket_upgrade_integration() -> Result<()> {
+        let url = crate::test::socket_io_server();
+        let socket = ClientBuilder::new(url)
+            .transport_type(TransportType::WebsocketUpgrade)
+            .connect_raw()
+            .await?;
+        test_socketio_socket(socket, "/".to_owned()).await
+    }
+
+    #[tokio::test]
+    async fn socket_io_any_integration() -> Result<()> {
+        let url = crate::test::socket_io_server();
+        let socket = ClientBuilder::new(url)
+            .transport_type(TransportType::Any)
+            .connect_raw()
+            .await?;
+        test_socketio_socket(socket, "/".to_owned()).await
+    }
+
+    async fn test_socketio_socket(socket: RawClient, nsp: String) -> Result<()> {
+        // open packet
+        let mut socket_stream = socket.as_stream();
+        let _: Option<Packet> = Some(socket_stream.next().await.unwrap()?);
+
+        let packet: Option<Packet> = Some(socket_stream.next().await.unwrap()?);
+
+        assert!(packet.is_some());
+
+        let packet = packet.unwrap();
+
+        assert_eq!(
+            packet,
+            Packet::new(
+                PacketId::Event,
+                nsp.clone(),
+                Some("[\"Hello from the message event!\"]".to_owned()),
+                None,
+                0,
+                None,
+            )
+        );
+
+        let packet: Option<Packet> = Some(socket_stream.next().await.unwrap()?);
+
+        assert!(packet.is_some());
+
+        let packet = packet.unwrap();
+
+        assert_eq!(
+            packet,
+            Packet::new(
+                PacketId::Event,
+                nsp.clone(),
+                Some("[\"test\",\"Hello from the test event!\"]".to_owned()),
+                None,
+                0,
+                None
+            )
+        );
+        let packet: Option<Packet> = Some(socket_stream.next().await.unwrap()?);
+
+        assert!(packet.is_some());
+
+        let packet = packet.unwrap();
+        assert_eq!(
+            packet,
+            Packet::new(
+                PacketId::BinaryEvent,
+                nsp.clone(),
+                None,
+                None,
+                1,
+                Some(vec![Bytes::from_static(&[4, 5, 6])]),
+            )
+        );
+
+        let packet: Option<Packet> = Some(socket_stream.next().await.unwrap()?);
+
+        assert!(packet.is_some());
+
+        let packet = packet.unwrap();
+        assert_eq!(
+            packet,
+            Packet::new(
+                PacketId::BinaryEvent,
+                nsp.clone(),
+                Some("\"test\"".to_owned()),
+                None,
+                1,
+                Some(vec![Bytes::from_static(&[1, 2, 3])]),
+            )
+        );
+
+        let cb = |message: Payload, _| {
+            async {
+                println!("Yehaa! My ack got acked?");
+                if let Payload::String(str) = message {
+                    println!("Received string ack");
+                    println!("Ack data: {}", str);
+                }
+            }
+            .boxed()
+        };
+
+        assert!(socket
+            .emit_with_ack(
+                "test",
+                Payload::String("123".to_owned()),
+                Duration::from_secs(10),
+                cb
+            )
+            .await
+            .is_ok());
+
+        let packet: Option<Packet> = Some(socket_stream.next().await.unwrap()?);
+
+        assert!(packet.is_some());
+        let packet = packet.unwrap();
+        assert_eq!(
+            packet,
+            Packet::new(
+                PacketId::Event,
+                nsp.clone(),
+                Some("[\"test-received\",123]".to_owned()),
+                None,
+                0,
+                None,
+            )
+        );
+
+        let packet: Option<Packet> = Some(socket_stream.next().await.unwrap()?);
+
+        assert!(packet.is_some());
+        let packet = packet.unwrap();
+        assert!(matches!(
+            packet,
+            Packet {
+                packet_type: PacketId::Ack,
+                nsp: _,
+                data: Some(_),
+                id: Some(_),
+                attachment_count: 0,
+                attachments: None,
+            }
+        ));
+
+        Ok(())
+    }
+}
