+++ socketio/src/asynchronous/mod.rs
@@ -5,3 +5,4 @@ mod socket;
 #[cfg(feature = "async")]
 pub use client::builder::ClientBuilder;
 pub use client::client::Client;
+pub use client::RawClient;
